+++ engine/runtime/src/main/java/org/enso/interpreter/node/expression/builtin/number/decimal/ToFloatNode.java
@@ -2,8 +2,8 @@ package org.enso.interpreter.node.expression.builtin.number.decimal;
 
 import org.enso.interpreter.dsl.BuiltinMethod;
 
-@BuiltinMethod(type = "Decimal", name = "to_decimal", description = "Identity on decimals")
-public class ToDecimalNode extends FloatNode {
+@BuiltinMethod(type = "Float", name = "to_float", description = "Identity on floats")
+public class ToFloatNode extends FloatNode {
   double execute(double self) {
     return self;
   }