+++ engine/runtime/src/main/java/org/enso/interpreter/node/expression/builtin/number/integer/ToFloatNode.java
@@ -8,13 +8,13 @@ import org.enso.interpreter.runtime.number.EnsoBigInteger;
 
 @BuiltinMethod(
     type = "Integer",
-    name = "to_decimal",
-    description = "Conversion of integers to decimals.")
-public abstract class ToDecimalNode extends IntegerNode {
+    name = "to_float",
+    description = "Conversion of integers to floats.")
+public abstract class ToFloatNode extends IntegerNode {
   public abstract Object execute(Object self);
 
-  public static ToDecimalNode build() {
-    return ToDecimalNodeGen.create();
+  public static ToFloatNode build() {
+    return ToFloatNodeGen.create();
   }
 
   @Specialization