+++ CHANGELOG.md
@@ -573,6 +573,7 @@
   `Column.map` and `Column.zip`.][7637]
 - [Added `delete_rows` method to Database Table, changed the
   `update_database_table` API into `update_rows`.][7709]
+- [Renamed `Decimal` to `Float`.][7807]
 
 [debug-shortcuts]:
   https://github.com/enso-org/enso/blob/develop/app/gui/docs/product/shortcuts.md#debug
@@ -811,6 +812,7 @@
 [7615]: https://github.com/enso-org/enso/pull/7615
 [7637]: https://github.com/enso-org/enso/pull/7637
 [7709]: https://github.com/enso-org/enso/pull/7709
+[7807]: https://github.com/enso-org/enso/pull/7807
 
 #### Enso Compiler